+++ src/template/HISTORY.md
@@ -1,6 +1,13 @@
 Template Change History
 =======================
 
+3.8.1
+-----
+
+* Added the concept of default options to `Y.Template` via the following
+  constructor signature: `Y.Template(engine, defaults)`.
+
+
 3.8.0
 -----